backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.2",
    "asyncpg>=0.31.0",
    "fastapi>=0.128.0",
    "passlib[bcrypt]>=1.7.4",
    "psycopg[binary]>=3.3.2",
    "pydantic-settings>=2.12.0",
    "python-jose>=3.5.0",
    "sqlmodel>=0.0.31",
    "uvicorn>=0.40.0",
]
